DynamoDB Query - fix bug in invalid operator detection
